Rename the `mocks` utility folder to `mocking`
The mocking utilities aren't mocks themselves, they're intended
to help build mocks.
