+++ src/core/__tests__/QueryManager/recycler.ts
@@ -10,7 +10,7 @@ import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
 import { stripSymbols } from '../../../utilities/testing/stripSymbols';
 import {
   MockSubscriptionLink
-} from '../../../utilities/testing/mocks/mockSubscriptionLink';
+} from '../../../utilities/testing/mocking/mockSubscriptionLink';
 
 // core
 import { QueryManager } from '../../QueryManager';
