+++ src/react/testing/index.ts
@@ -1,9 +1,9 @@
-export { MockedProvider } from '../../utilities/testing/mocks/MockedProvider';
-export { MockLink, mockSingleLink } from '../../utilities/testing/mocks/mockLink';
+export { MockedProvider } from '../../utilities/testing/mocking/MockedProvider';
+export { MockLink, mockSingleLink } from '../../utilities/testing/mocking/mockLink';
 export {
   MockSubscriptionLink,
   mockObservableLink
-} from '../../utilities/testing/mocks/mockSubscriptionLink';
+} from '../../utilities/testing/mocking/mockSubscriptionLink';
 
-export { createMockClient } from '../../utilities/testing/mocks/mockClient';
+export { createMockClient } from '../../utilities/testing/mocking/mockClient';
 export { stripSymbols } from '../../utilities/testing/stripSymbols';
