+++ src/utilities/testing/mocking/mockClient.ts
@@ -3,7 +3,7 @@ import { DocumentNode } from 'graphql';
 import { ApolloClient } from '../../../ApolloClient';
 import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
 import { NormalizedCacheObject } from '../../../cache/inmemory/types';
-import { mockSingleLink } from '../../../utilities/testing/mocks/mockLink';
+import { mockSingleLink } from '../../../utilities/testing/mocking/mockLink';
 
 export function createMockClient<TData>(
   data: TData,
